# Improve travis.yml config for faster CI