--- a/ci/download_bazel.sh
+++ b/ci/download_bazel.sh
@@ -1,0 +1,49 @@
+#!/bin/sh
+# Copyright 2019 The TensorFlow Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+
+# Script to download Bazel binary directly onto a build machine.
+
+set -e
+
+die() {
+  printf >&2 "%s\n" "$1"
+  exit 1
+}
+
+if [ "$#" -ne 3 ]; then
+  die "Usage: ${0} <version> <sha256sum> <destination-file>"
+fi
+
+version="$1"
+checksum="$2"
+dest="$3"
+
+temp_dest="$(mktemp)"
+
+mirror_url="https://mirror.bazel.build/github.com/bazelbuild/bazel/releases/download/${version}/bazel-${version}-linux-x86_64"
+github_url="https://github.com/bazelbuild/bazel/releases/download/${version}/bazel-${version}-linux-x86_64"
+
+for url in "${mirror_url}" "${github_url}"; do
+  wget -t 3 -O "${temp_dest}" "${url}" \
+    && printf "%s  %s\n" "${checksum}" "${temp_dest}" | shasum -a 256 --check \
+    || { rm -f "${temp_dest}"; continue; }
+  mv "${temp_dest}" "${dest}"
+  break
+done
+
+[ -f "${dest}" ]
+chmod +x "${dest}"
+ls -l "${dest}"