This streamlines and improves our .travis.yml configuration and caching behavior in hopes of reducing our painfully slow Travis builds.  In my testing so far this drops the total time to execute by ~30% (so from 50m total across 3 machines to more like 35m), by shaving a little over 2 minutes off of each of the 6 jobs in our test matrix.

Here's the list of changes I implemented that I felt overall improved speed (or were mixed, but seemed like better practice):
- changed caching approach to use Bazel's properly designed caching APIs
  - repository caching for `bazel fetch` results
  - remote caching API for build results/artifacts, using local disk backend
  - **con:** fetch is slower (60ish seconds instead of 15) - not entirely sure why, but perhaps it doesn't cache all the downloaded results of fetch (some indication of this on https://github.com/bazelbuild/bazel/issues/6869). We may be able to tweak fetching behavior to improve this.
  - **con:** build is slower (90ish seconds instead of 60) - I think with the new caching we're rebuilding more things, e.g. now there are a bunch of closure warnings. We may be able to improve this by restructuring some of our build rules to be more cacheable, or generally speeding up the build.
  - **pro:** fetch/store operations for cache are much faster (60s fetch + 140s store reduced to 15s fetch + 60-75s store), which basically makes up for the above slowdowns
  - **pro:** no more need for hacky code to manually "prune" the output directory
- added caching for pip packages
  - **pro:** pip installs are perhaps 30% or so faster, which also speeds up the smoke test (I'm not sure why they aren't even faster than this; I think sometimes incorrectly skip the cache or in some cases the packages run setup.py operations that still take time)
  - **pro:** less at the whim of pypi latency; less uninteresting logging of package downloads
  - **con:** some extra time to fetch/store cache but not much (maybe 15s)
- removed warning-only flake8 run
  - **pro:** 10 seconds saved for every build job
  - **con:** no notice of warnings, but we never looked at this routinely anyway
- changed bazel --action_env=PATH to apply to build+test, not just test
  - **pro:** can re-use package analysis from build step in test step, saving a few seconds

And then I did some cleanup:

- factored out bazel settings into a `ci/bazelrc` file
- factored out bazel downloading logic into `ci/download_bazel.sh` and streamlined it slightly
- added bazelrc `common --curses=no` to suppress attempted progress bar spam
- removed aggressive use of bazel --host_jvm_args that I didn't see justification for
- removed redundant travis.yml config elements like sudo and os (Fixes #2113)
- slightly reworked some other misc travis steps for clarity

Finally I added some simple instrumentation showing the elapsed time since the shell started at various points during the job, which makes it a bit easier to get a sense of where the slow parts are without having to sum up the time taken by many individual commands (also, the travis time indicators are slightly flaky and don't always get parsed correctly by the fancy UI).

-----

I was originally hoping for a bigger speedup by combining all the testing for the different TF versions into a single job, since then we could reuse the build step and avoid rebuilding each time.  The problem I ran into was that bazel test result caching falls apart at that point - I wasn't aware of a simple way to tell bazel to treat each run of the tests as a separate set of cacheable test results (rather than re-using the same cache, which due to the non-hermeticity could fail to rerun tests that need to be rerun).  Trying to work around this via cache munging would erase a lot of the benefits of the simplification above, and that approach was already much more complicated than what I have here.  

Down the road, I think further improvements could be:
- making the actual fetch and build steps faster / caching them more effectively
- figuring out a way to incorporate the TF version into Bazel's hermeticized understanding of the world so that it caches test results for different TF versions differently, but somehow still re-uses the build outputs (I have no idea how to do this but maybe we could ask the Bazel team)
- reducing our TF-dependent tests to a smaller set, and then only running a much more limited set of tests against each TF version we test against (though of course once we bump versions to 2.0 we could stop testing against TF 1.x)
- waiting for Python 2 to finally die in 2020 and then no longer running those tests either