@nfelt 2019-05-25
For comparison of build performance, here's a job with the existing Travis config that ran essentially no tests (typo fix PR) and took a total of 49 minutes: https://travis-ci.com/tensorflow/tensorboard/builds/113105919

With this PR, running with an empty commit (so the cache is warm, and also running no tests) it takes a total of 37.5 minutes: https://travis-ci.com/tensorflow/tensorboard/builds/113164615